fastapi2/agent/requirements.txt
# Chiranjeevi Medical Agent — Dependencies
# Install:  pip install -r agent/requirements.txt

langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
tavily-python>=0.5.0
llama-cpp-python>=0.2.90
xmltodict>=0.13.0
python-dotenv>=1.0.1
requests>=2.32.0
pydantic>=2.10.0
